// src/types/writer-persistence-authorization-remediation-review-no-go.ts
export type WriterPersistenceAuthorizationRemediationReviewNoGoMode =
  "persistence_adapter_implementation_authorization_remediation_review_no_go_packet_only";

export type WriterPersistenceAuthorizationRemediationReviewNoGoStatus =
  | "no_go"
  | "manual_review_blocked";

export type WriterPersistenceAuthorizationRemediationReviewNoGoItem = {
  id: string;
  category: WriterPersistenceAuthorizationRemediationCategory;
  title: string;
  status: WriterPersistenceAuthorizationRemediationReviewNoGoStatus;
  owner: WriterPersistenceAuthorizationRemediationOwner;
  sourceReviewItemIds: string[];
  sourceRemediationItemIds: string[];
  sourceNoGoItemIds: string[];
  sourceRefs: string[];
  noGoQuestion: string;
  noGoConclusion: string;
  blockingEvidence: string[];
  unresolvedReviewGaps: string[];
  forbiddenShortcuts: string[];
  reconsiderationRequirements: string[];
  safeEscalationRefs: string[];
  redactionRules: string[];
  nonAcceptanceClauses: string[];
  nextSafeAction: string;
};

export type WriterPersistenceAuthorizationRemediationReviewNoGoRuntimeFlags = {
  allRuntimeEffectsBlocked: true;
  wouldAcceptRemediationReviewNoGo: false;
  wouldRecordRemediationReviewNoGo: false;
  wouldDenyImplementationAuthorizationFromReview: false;
  wouldPromoteToAuthorizationReconsideration: false;
  wouldAcceptRemediationReview: false;
  wouldRecordRemediationReview: false;
  wouldStoreRemediationReviewEvidence: false;
  wouldMarkExternalRemediationReviewed: false;
  wouldAcceptExternalRemediationState: false;
  wouldAcceptExternalApprovalArchive: false;
  wouldStoreApprovalArtifact: false;
  wouldUploadApprovalArtifact: false;
  wouldReadExternalArtifact: false;
  wouldHashExternalArtifact: false;
  wouldPersistArchiveIndex: false;
  wouldMarkArchiveComplete: false;
  wouldCreateAuthorizationRecord: false;
  wouldRecordAuthorizationDecision: false;
  wouldRecordAuthorizationNoGoDecision: false;
  wouldAcceptAuthorizationNoGoDecision: false;
  wouldAcceptRemediationPlan: false;
  wouldRecordRemediationEvidence: false;
  wouldMarkBlockerResolved: false;
  wouldCreateRemediationTicket: false;
  wouldDenyImplementationAuthorization: false;
  wouldGrantImplementationAuthorization: false;
  wouldRecordHumanDecision: false;
  wouldAcceptHumanDecision: false;
  wouldStoreDecisionArtifact: false;
  wouldAcceptReleaseNoGo: false;
  wouldRecordGoDecision: false;
  wouldGrantReleaseApproval: false;
  wouldEnableFeatureFlag: false;
  wouldDeployCode: false;
  wouldRunProductionWriter: false;
  wouldCollectSignature: false;
  wouldRecordOwnerApproval: false;
  wouldGrantImplementationApproval: false;
  wouldCreateApprovalRecord: false;
  wouldAcceptPatchReview: false;
  wouldReviewRealPatch: false;
  wouldAcceptPatch: false;
  wouldGeneratePatch: false;
  wouldApplyPatch: false;
  wouldModifyFiles: false;
  wouldCreateFiles: false;
  wouldDeleteFiles: false;
  wouldRunGitCommand: false;
  wouldCreateBranch: false;
  wouldCheckoutBranch: false;
  wouldCreatePullRequest: false;
  wouldCreateTestFiles: false;
  wouldRunAutomatedTests: false;
  wouldCreateImplementationPlan: false;
  wouldCreateImplementationBranch: false;
  wouldCreateAdapterCode: false;
  wouldImportRealWriterImplementation: false;
  wouldRunTransaction: false;
  wouldCreateServiceRoleClient: false;
  wouldReadServiceRoleSecret: false;
  wouldPersistEvidence: false;
  wouldStoreRawPayload: false;
  wouldStoreSecrets: false;
  wouldWriteRows: false;
  wouldWriteAuditRows: false;
  wouldReserveIdempotencyKeys: false;
  wouldWriteIdempotencyRows: false;
  wouldWriteCompensationRows: false;
  wouldCreateMigrationFile: false;
  wouldApplyMigration: false;
  wouldCreateTables: false;
  wouldEnableWriters: false;
  wouldCallAi: false;
  wouldCallStripe: false;
  wouldUnlockReports: false;
};

export type WriterPersistenceAuthorizationRemediationReviewNoGoPayload =
  WriterPersistenceAuthorizationRemediationReviewNoGoRuntimeFlags & {
    safeMode: true;
    readOnly: true;
    remediationReviewNoGoMode: WriterPersistenceAuthorizationRemediationReviewNoGoMode;
    sourceRemediationReviewChecklistMode: WriterPersistenceAuthorizationRemediationReviewMode;
    checkedAt: string;
    noGoItemCount: number;
    noGoCount: number;
    manualReviewBlockedCount: number;
    reconsiderationStillBlockedCount: number;
    blockingEvidenceCount: number;
    unresolvedReviewGapCount: number;
    forbiddenShortcutCount: number;
    reconsiderationRequirementCount: number;
    sourceReviewItemCount: number;
    sourceExternalEvidenceMissingCount: number;
    sourceManualReviewerRequiredCount: number;
    reviewNoGoPacketReady: true;
    reviewNoGoPacketOnly: true;
    sourceReviewChecklistReady: true;
    sourceReviewChecklistOnly: true;
    sourceRemediationPlanReady: true;
    sourceRemediationPlanOnly: true;
    sourceReleaseStillBlocked: true;
    externalRemediationStatesAccepted: false;
    remediationReviewAccepted: false;
    remediationReviewComplete: false;
    remediationReviewNoGoAccepted: false;
    remediationReviewNoGoRecorded: false;
    implementationAuthorizationReconsiderationReady: false;
    implementationAuthorizationRemediationAccepted: false;
    implementationAuthorizationDecisionReady: false;
    implementationAuthorizationDecisionRecorded: false;
    implementationAuthorizationNoGoAccepted: false;
    implementationAuthorizationDenied: false;
    implementationAuthorizationGranted: false;
    implementationAuthorized: false;
    authorizationDecisionRecorded: false;
    authorizationArtifactStored: false;
    externalApprovalArchiveAccepted: false;
    archiveCompletenessAccepted: false;
    implementationApprovalGranted: false;
    implementationBranchApproved: false;
    implementationPlanApproved: false;
    readyToApplyPatch: false;
    readyToCreateImplementationBranch: false;
    readyForAdapterImplementation: false;
    readyForReleaseExecution: false;
    adapterImplemented: false;
    adapterImplementationApproved: false;
    adapterImplementationAllowed: false;
    implementationReviewComplete: false;
    allOwnerApprovalsComplete: false;
    allBlockingEvidenceReady: false;
    blockedCodes: string[];
    reviewNoGoRules: string[];
    reconsiderationRules: string[];
    sourceBlockedCodes: string[];
    noGoItems: WriterPersistenceAuthorizationRemediationReviewNoGoItem[];
  };

export type WriterPersistenceAuthorizationRemediationReviewNoGoProbeResult =
  WriterPersistenceAuthorizationRemediationReviewNoGoRuntimeFlags & {
    safeMode: true;
    readOnly: true;
    blocked: true;
    remediationReviewNoGoMode: WriterPersistenceAuthorizationRemediationReviewNoGoMode;
    itemId?: string;
    itemTitle?: string;
    itemStatus?: WriterPersistenceAuthorizationRemediationReviewNoGoStatus;
    summary: string;
    reviewNoGoPacketOnly: true;
    sourceReleaseStillBlocked: true;
    externalRemediationStatesAccepted: false;
    remediationReviewAccepted: false;
    remediationReviewComplete: false;
    remediationReviewNoGoAccepted: false;
    remediationReviewNoGoRecorded: false;
    implementationAuthorizationReconsiderationReady: false;
    implementationAuthorizationGranted: false;
    implementationAuthorized: false;
    authorizationDecisionRecorded: false;
    authorizationArtifactStored: false;
    externalApprovalArchiveAccepted: false;
    archiveCompletenessAccepted: false;
    readyToCreateImplementationBranch: false;
    readyForAdapterImplementation: false;
    readyForReleaseExecution: false;
    adapterImplemented: false;
    adapterImplementationApproved: false;
    adapterImplementationAllowed: false;
    allOwnerApprovalsComplete: false;
    allBlockingEvidenceReady: false;
    blockedCodes: string[];
    noGoItems: WriterPersistenceAuthorizationRemediationReviewNoGoItem[];
  };